app_requirements.txt
pendulum
pint
attrs
pygeodesy
sqlalchemy
pika
appdirs
accessible_output2
https://github.com/tyrylu/uimanager/archive/master.zip
blinker
PySide2
anglr
bitmath
pyopenal
pyogg
shapely
jinja2
requests